requirements/base.txt
psycopg2==2.5
Pillow==2.0.0
Django==1.5.1
South==0.8.1
six==1.3.0
BeautifulSoup==3.2.1
pycsw==1.6.0
sorl-thumbnail==11.12
argparse==1.2.1
pytz==2012c
pysolr==3.0.6
requests==1.2.3

django_compressor==1.3
django-appconf==0.6
django-ratings==0.3.7
django-pagination==1.0.7
django-sorting==0.1
django-widget-tweaks==1.3
django-haystack==2.0.0
django-selectable==0.7.0
